+++ rpc/src/endpoint/blockchain.rs
@@ -41,6 +41,8 @@ impl crate::Request for Request {
     }
 }
 
+impl crate::SimpleRequest for Request {}
+
 /// Block responses
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct Response {